! fix #3569 RSS feeds were broken
